# src/Services/FirmwareService/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/firmware/v1/ota/check": {
            "get": {
                "description": "Ищет активную кампанию обновления для переданного контроллера. Возвращает ссылку на скачивание бинарника из S3.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Device API"
                ],
                "summary": "Проверка обновлений (OTA)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID контроллера",
                        "name": "controller_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Возвращает статус обновления и метаданные",
                        "schema": {
                            "$ref": "#/definitions/internal_presentation.OTAResponse"
                        }
                    },
                    "400": {
                        "description": "Неверный формат controller_id",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/firmware/v1/ota/status": {
            "post": {
                "description": "Вызывается ESP32 для сообщения о начале скачивания, успешной прошивке или ошибке обновления.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Device API"
                ],
                "summary": "Обновить статус прошивки устройства",
                "parameters": [
                    {
                        "description": "Данные о статусе обновления",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_presentation.UpdateTargetStatusRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Статус успешно обновлен"
                    },
                    "400": {
                        "description": "Неверный формат запроса",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "internal_presentation.OTAResponse": {
            "type": "object",
            "properties": {
                "download_url": {
                    "type": "string"
                },
                "file_hash": {
                    "type": "string"
                },
                "has_update": {
                    "type": "boolean"
                },
                "version": {
                    "type": "string"
                }
            }
        },
...
        "internal_presentation.UpdateTargetStatusRequest": {
            "type": "object",
            "properties": {
                "campaign_id": {
                    "type": "string"
                },
                "controller_id": {
                    "type": "string"
                },
                "error_message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        }
    }
}